server/requirements.txt
alembic==1.4.3
click==7.1.2
jsonpatch==1.32
jsonpointer==2.1
loguru==0.5.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.24.2
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.37.4
djangorestframework==3.12.0
Django==2.2.0
django-filter
Pygments==2.7.4
Markdown==3.2.1
coreapi==2.3.3
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.0.1
PyYAML==5.4
django-cors-headers==3.7.0
djangorestframework-link-header-pagination
djangorestframework-simplejwt